bugfix favourite action for all dashboard blocks
