+++ Modules/LearningSequence/classes/class.ilDashboardLearningSequenceGUI.php
@@ -111,16 +111,6 @@ class ilDashboardLearningSequenceGUI extends ilDashboardBlockGUI
         return 'pdlern';
     }
 
-    public function addCustomCommandsToActionMenu(ilObjectListGUI $itemListGui, int $ref_id): void
-    {
-        $this->ctrl->setParameter($this, "item_ref_id", $ref_id);
-        $itemListGui->addCustomCommand(
-            $this->ctrl->getLinkTarget($this, "addToDesk"),
-            "rep_add_to_favourites"
-        );
-        $this->ctrl->clearParameterByClass(self::class, "item_ref_id");
-    }
-
     public function confirmedRemoveObject(): void
     {
         $refIds = (array) ($this->http->request()->getParsedBody()['ref_id'] ?? []);
